feat(ConnectedButtonProfile): moved to separate component
