+++ packages/epics/src/profile/components/index.ts
@@ -1,1 +1,2 @@
 export * from './button-profile';
+export * from './connected-button-profile';
